[service] Move Software to Software::Manager
