+++ service/test/dinstaller/software/manager_test.rb
@@ -19,11 +19,11 @@
 # To contact SUSE LLC about this file by physical or electronic mail, you may
 # find current contact information at www.suse.com.
 
-require_relative "../test_helper"
+require_relative "../../test_helper"
 require "dinstaller/config"
-require "dinstaller/software"
+require "dinstaller/software/manager"
 
-describe DInstaller::Software do
+describe DInstaller::Software::Manager do
   subject { described_class.new(config, logger) }
 
   let(:logger) { Logger.new($stdout, level: :warn) }
@@ -56,8 +56,8 @@ describe DInstaller::Software do
     let(:backup_repos_dir) { File.join(rootdir, "etc", "zypp", "repos.d.backup") }
 
     before do
-      stub_const("DInstaller::Software::REPOS_DIR", repos_dir)
-      stub_const("DInstaller::Software::REPOS_BACKUP", backup_repos_dir)
+      stub_const("DInstaller::Software::Manager::REPOS_DIR", repos_dir)
+      stub_const("DInstaller::Software::Manager::REPOS_BACKUP", backup_repos_dir)
       FileUtils.mkdir_p(repos_dir)
     end
 
@@ -185,8 +185,8 @@ describe DInstaller::Software do
     let(:backup_repos_dir) { File.join(rootdir, "etc", "zypp", "repos.d.backup") }
 
     before do
-      stub_const("DInstaller::Software::REPOS_DIR", repos_dir)
-      stub_const("DInstaller::Software::REPOS_BACKUP", backup_repos_dir)
+      stub_const("DInstaller::Software::Manager::REPOS_DIR", repos_dir)
+      stub_const("DInstaller::Software::Manager::REPOS_BACKUP", backup_repos_dir)
       FileUtils.mkdir_p(repos_dir)
       FileUtils.mkdir_p(backup_repos_dir)
       FileUtils.touch(File.join(backup_repos_dir, "example.repo"))
